\name{usutuFrance-package}
\alias{usutuFrance-package}
\alias{usutuFrance}
\docType{package}
\title{
\packageTitle{usutuFrance}
}
\description{
\packageDescription{usutuFrance}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{usutuFrance}
\packageIndices{usutuFrance}
}
\author{
\packageAuthor{usutuFrance}

Maintainer: \packageMaintainer{usutuFrance}
}
\references{
  Bouchez-Zacria, M., Calenge, C., Villers, A., Lecollinet, S.,
  Gonzalez, G., Quintard, B., Leclerc, A., Baurier, F., Paty, M.-C.,
  Eraud, C. and Decors, A. 2024. Analysis of the Usutu episode of summer
  2018 in birds in France. In prep.
}
\keyword{ package }
\examples{
\dontrun{
vignette("usutuFrance")
}
}
